[BTAT-132] Refactored the Feedback into its own package
