+++ app/views/helpers/phaseBanner.scala.html
@@ -19,6 +19,6 @@
 <div class="phase-banner">
   <p>
     <strong id="phase" class="phase-tag">@phaseId</strong>
-    <span>This is a new service – your <a href="@controllers.routes.FeedbackController.show()">feedback</a> will help us to improve it.</span>
+    <span>This is a new service – your <a href="@controllers.feedback.routes.FeedbackController.show()">feedback</a> will help us to improve it.</span>
   </p>
 </div>
